Time: 0 ms (100%), Space: 8.5 MB (66.11%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -7,7 +7,7 @@ public:
             return false;
         }
         
-        long tempx = (long) x;
+        long tempx = x;
         long num = 0;
         int place = 1;
